+++ src/aux-common/runtime/AuxLibrary.spec.ts
@@ -157,6 +157,7 @@ import {
     enableCustomDragging,
     registerCustomApp,
     setAppOutput,
+    unregisterCustomApp,
 } from '../bots';
 import { types } from 'util';
 import {
@@ -192,6 +193,9 @@ import {
 import { RanOutOfEnergyError } from './AuxResults';
 import { Subscription, SubscriptionLike } from 'rxjs';
 import { waitAsync } from '../test/TestHelpers';
+import { embedBase64InPdf } from './Utils';
+import { fromByteArray, toByteArray } from 'base64-js';
+import { Fragment } from 'preact';
 
 const uuidMock: jest.Mock = <any>uuid;
 jest.mock('uuid');
@@ -2147,6 +2151,49 @@ describe('AuxLibrary', () => {
                     'test.aux',
                     'application/json'
                 );
+
+                expect(action).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should create a PDF if the .pdf extension is used', () => {
+                const action = library.api.os.downloadBots(
+                    [bot1, bot2],
+                    'test.pdf'
+                );
+                const json = JSON.stringify({
+                    version: 1,
+                    state: {
+                        [bot1.id]: bot1,
+                        [bot2.id]: bot2,
+                    },
+                });
+                const encoder = new TextEncoder();
+                const bytes = encoder.encode(json);
+                const base64 = fromByteArray(bytes);
+
+                const expected = download(
+                    embedBase64InPdf(base64),
+                    'test.pdf',
+                    'application/pdf'
+                );
+
+                expect(action).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should create PDFs that can be parsed', () => {
+                const action = library.api.os.downloadBots(
+                    [bot1, bot2],
+                    'test.pdf'
+                );
+
+                const bots = library.api.os.parseBotsFromData(action.data);
+
+                expect(bots).toEqual([
+                    createBot(bot1.id, bot1.tags),
+                    createBot(bot2.id, bot2.tags),
+                ]);
             });
         });
 
@@ -2358,6 +2405,74 @@ describe('AuxLibrary', () => {
                 const action = library.api.os.importAUX(json);
                 expect(action).toEqual(addState(uploadState));
                 expect(context.actions).toEqual([addState(uploadState)]);
+            });
+
+            it('should be able to parse PDF files', () => {
+                const uploadState: BotsState = {
+                    [bot1.id]: createBot(bot1.id, bot1.tags),
+                };
+                const downloadAction = library.api.os.downloadBots(
+                    [bot1],
+                    'test.pdf'
+                );
+                const action = library.api.os.importAUX(downloadAction.data);
+                expect(action).toEqual(addState(uploadState));
+                expect(context.actions.slice(1)).toEqual([
+                    addState(uploadState),
+                ]);
+            });
+        });
+
+        describe('os.parseBotsFromData()', () => {
+            it('should return the list of bots that are in the given JSON', () => {
+                const json = JSON.stringify({
+                    version: 1,
+                    state: {
+                        [bot1.id]: bot1,
+                        [bot2.id]: bot2,
+                    },
+                });
+
+                const bots = library.api.os.parseBotsFromData(json);
+
+                expect(bots).toEqual([
+                    createBot(bot1.id, bot1.tags),
+                    createBot(bot2.id, bot2.tags),
+                ]);
+            });
+
+            it('should return the list of bots that are in the given PDF', () => {
+                const json = JSON.stringify({
+                    version: 1,
+                    state: {
+                        [bot1.id]: bot1,
+                        [bot2.id]: bot2,
+                    },
+                });
+                const encoder = new TextEncoder();
+                const bytes = encoder.encode(json);
+                const base64 = fromByteArray(bytes);
+                const pdf = embedBase64InPdf(base64);
+
+                const bots = library.api.os.parseBotsFromData(pdf);
+
+                expect(bots).toEqual([
+                    createBot(bot1.id, bot1.tags),
+                    createBot(bot2.id, bot2.tags),
+                ]);
+            });
+
+            it('should return null if the data is not JSON or a PDF', () => {
+                const bots = library.api.os.parseBotsFromData('abcdef');
+                expect(bots).toEqual(null);
+            });
+
+            it('should return null if the data in the PDF is not bots', () => {
+                const pdf = embedBase64InPdf('abcdefghijfk');
+
+                const bots = library.api.os.parseBotsFromData(pdf);
+
+                expect(bots).toEqual(null);
             });
         });
 
@@ -3263,6 +3378,18 @@ describe('AuxLibrary', () => {
                 const expected = registerCustomApp(
                     'testPortal',
                     bot1.id,
+                    context.tasks.size
+                );
+                expect(promise[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+        });
+
+        describe('os.unregisterApp()', () => {
+            it('should return a UnregisterCustomPortal action', () => {
+                const promise: any = library.api.os.unregisterApp('testPortal');
+                const expected = unregisterCustomApp(
+                    'testPortal',
                     context.tasks.size
                 );
                 expect(promise[ORIGINAL_OBJECT]).toEqual(expected);
@@ -12064,6 +12191,19 @@ describe('AuxLibrary', () => {
             `;
 
             expect(result).toMatchSnapshot();
+        });
+
+        describe('h()', () => {
+            it('should return a HTML VDOM element', () => {
+                const result = library.api.html.h('h1', null, 'Hello, World!');
+                expect(result).toMatchSnapshot();
+            });
+        });
+
+        describe('f', () => {
+            it('should be the Fragment element type', () => {
+                expect(library.api.html.f).toBe(Fragment);
+            });
         });
     });
 });