+++ src/aux-server/aux-web/shared/MonacoHelpers.ts
@@ -65,8 +65,7 @@ import { CurrentVersion } from '@casual-simulation/causal-trees';
 import { Color } from '@casual-simulation/three';
 import { invertColor } from './scene/ColorUtils';
 import { getCursorColorClass, getCursorLabelClass } from './StyleHelpers';
-// import jscodeshift from 'jscodeshift';
-// import MonacoJSXHighlighter from './public/monaco-jsx-highlighter/index';
+import MonacoJSXHighlighter from './public/monaco-jsx-highlighter/index';
 import axios from 'axios';
 import { customPortalLanguageId } from './public/monaco-editor/custom-portal-typescript/custom-portal-typescript.contribution';
 import {
@@ -394,11 +393,10 @@ export function watchEditor(
     simulation: Simulation,
     editor: monaco.editor.ICodeEditor
 ): Subscription {
-    // const monacoJsxHighlighter = new MonacoJSXHighlighter(
-    //     monaco,
-    //     jscodeshift,
-    //     editor
-    // );
+    const monacoJsxHighlighter = new MonacoJSXHighlighter(
+        new Transpiler(),
+        editor
+    );
 
     const modelChangeObservable = new Observable<
         monaco.editor.IModelChangedEvent
@@ -591,35 +589,35 @@ export function watchEditor(
         )
     );
 
-    // const enableJsxHighlightingOnCorrectModels = modelInfos.pipe(
-    //     map(
-    //         (info) =>
-    //             info.language === 'javascript' || info.language === 'typescript'
-    //     ),
-    //     scan(
-    //         (acc, needsJsxHighlighting) => {
-    //             if (acc) {
-    //                 acc();
-    //             }
-    //             if (needsJsxHighlighting) {
-    //                 return monacoJsxHighlighter.highLightOnDidChangeModelContent(
-    //                     undefined,
-    //                     () => {},
-    //                     undefined,
-    //                     () => {}
-    //                 );
-    //             } else {
-    //                 return () => {};
-    //             }
-    //         },
-    //         () => {}
-    //     )
-    // );
+    const enableJsxHighlightingOnCorrectModels = modelInfos.pipe(
+        map(
+            (info) =>
+                info.language === 'javascript' || info.language === 'typescript'
+        ),
+        scan(
+            (acc, needsJsxHighlighting) => {
+                if (acc) {
+                    acc();
+                }
+                if (needsJsxHighlighting) {
+                    return monacoJsxHighlighter.highLightOnDidChangeModelContent(
+                        undefined,
+                        () => {},
+                        undefined,
+                        (err) => console.error(err)
+                    );
+                } else {
+                    return () => {};
+                }
+            },
+            () => {}
+        )
+    );
 
     const sub = new Subscription();
 
     sub.add(decorators.subscribe());
-    // sub.add(enableJsxHighlightingOnCorrectModels.subscribe());
+    sub.add(enableJsxHighlightingOnCorrectModels.subscribe());
 
     sub.add(
         toSubscription(