+++ src/aux-vm/portals/CustomAppHelper.spec.ts
@@ -8,6 +8,7 @@ import {
     iteratePartitions,
     MemoryPartition,
     registerCustomApp,
+    unregisterCustomApp,
 } from '@casual-simulation/aux-common';
 import { Subscription } from 'rxjs';
 import { tap } from 'rxjs/operators';
@@ -107,6 +108,18 @@ describe('CustomAppHelper', () => {
 
                 expect(values[0]).toBeInstanceOf(HtmlAppBackend);
                 expect(values[0].botId).toBe(null);
+            });
+        });
+
+        describe('unregister_custom_app', () => {
+            it('should delete the portal for the given event', () => {
+                portals.handleEvents([registerCustomApp('htmlPortal', null)]);
+
+                expect([...portals.portals.keys()]).toEqual(['htmlPortal']);
+
+                portals.handleEvents([unregisterCustomApp('htmlPortal')]);
+
+                expect(portals.portals.size).toBe(0);
             });
         });
     });