+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # CasualOS Changelog
 
+## V2.0.4
+
+#### Date: 7/27/2021
+
+### :rocket: Improvements
+
+-   Added the ability to download a PDF with embedded bot data by specifying a filename with a `.pdf` extension to `os.downloadBots()`.
+-   Added the `os.parseBotsFromData(data)` function.
+    -   This function can parse a list of bot mods from JSON or from the contents of a PDF that was created with `os.downloadBots()`.
+    -   It returns a list of bot mods (i.e. mods that have the structure of bots) which can in turn be passed to `create()` to add them to the server.
+-   Added the `os.unregisterApp(appID)` function to allow removing apps after they have been registered.
+-   Added the ability to use [JSX](https://reactjs.org/docs/introducing-jsx.html) for Apps instead of the `html` string helper.
+    -   JSX allows you to use a HTML-like language directly inside listeners. This provides some nice benefits including proper syntax highlighting and error messages.
+    -   For example:
+        ```javascript
+        let result = <h1>Hello, World!</h1>;
+        ```
+    -   Due to convienience this will probably become the preferred way to write HTML for apps, however the `html` string helper will still be available.
+
 ## V2.0.3
 
 #### Date: 7/19/2021